@@ -17,7 +17,7 @@
         }
     ],
     "require": {
-        "php": "^7.1|^8",
+        "php": "^7.4|^8",
         "psr/log": "^1|^2|^3",
         "symfony/var-dumper": "^4|^5|^6|^7"
     },
